updates to Managers schedule
I made updates to the manager's schedule based on running the activity on the first day. I also added a project option for scheduling that is related but now has chemicals in a warehouse, that will cause them to at least look at a different situation and make the scheduling connection.